Move pager.xml parser into pager module
